% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CD4map_plot}
\alias{CD4map_plot}
\title{Plot your query scRNAseq data on top of the integrated helper T cell reference dataset}
\usage{
CD4map_plot(annotated_query, coloring_based_on = "cluster")
}
\arguments{
\item{annotated_query}{Seurat object annotated with CD4map::CD4map_annotate() function}

\item{coloring_based_on}{Allowed values: "cluster", "score". If set to "cluster" (default), color encodes predicted cell annotations; if set to "score", color encodes the prediction score}
}
\value{
ggplot object
}
\description{
Plot your query scRNAseq data on top of the integrated helper T cell reference dataset
}
